Client/requirements.txt
grpcio 
googleapis-common-protos 
redis
